+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add new “Select Enclosing Symbols” and “Split Selection by Lines” commands to the Edit > Select menu.
 - Add new “Show in Finder” command to the File menu.
 - Add new “Create Document in CotEditor” shortcut action.
+- Add the “Highlight” color to the CotEditor theme format.
 - Support the alpha channel for the current line in theme settings.
 - Add Assembly syntax.
 - Add new “Resinifictrix (Dark)” theme.
@@ -23,9 +24,10 @@
 - Change the unit of character ranges handled in CotEditor Scripting for AppleScript from UTF-16 based to the Unicode grapheme cluster-based (This is to follow the specification change in AppleScript 2.0 introduced in Mac OS X 10.5).
 - Improve VoiceOver support in the Quick Action bar.
 - Remove Solarized themes from the bundle.
-- Update all the bundled themes to have a 70% opacity in the current line highlight.
+- Update all the bundled themes to have the custom highlight color and a 70% opacity in the current line highlight.
 - Improve the performance of counting values in the editor for the status bar and the document inspector to avoid flicking.
 - Set the editor's kerning to zero when a monospaced font is set via the font panel.
+- Avoid loading errors on themes when missing value exists.
 - Make more table columns sortable.
 - [trivial] Organize the structure of the Edit menu.
 - [trivial] Suppress display of the “Extracting” message on the navigation bar in instantaneous parsing.